requirements.txt
flask
feedparser
requests
readability-lxml
langchain
langchain-community
beautifulsoup4
gunicorn
openai